+++ changelog.md
@@ -1,4 +1,11 @@
 # Changelog
 
-**./test-parts/test-graph-parse.js**
-* Complete
+**./test-parts/test-pathfinding-input.js**
+* New file - Unit tests for pathfinding input.
+* So far, only the 'exact route' cases have been implemented.
+
+---
+
+**./test/index.js**
+* Added requirement: '../test-parts/test-pathfinding-input'
+* Call 'testPathfindingInput'